+++ src/mongo/db/curop.h
@@ -156,8 +156,8 @@ public:
 
     // detailed options
     long long cursorid;
-    int ntoreturn;
-    int ntoskip;
+    long long ntoreturn;
+    long long ntoskip;
     bool exhaust;
 
     // debugging/profile info
@@ -187,7 +187,7 @@ public:
 
     // response info
     int executionTime;
-    int nreturned;
+    long long nreturned;
     int responseLength;
 };
